FAR patch 01803.BugZ1096_Macro2.0 Дата              : 15.06.2004 Сделал            : Valentin Skirdin Описание          : BugZ#1096 - shiftins в залоченном файле - двойное выделение                     Macro (ver 2.0 pre) Измененные файлы  : edit.cpp editor.cpp far.dep far.dsp Far.mak fn.hpp                     macro.hpp main.cpp mkfar.mak Новые файлы       : syntax.cpp syntax.hpp Удаленные файлы   : Состав            : 01803.BugZ1096.txt                     01803.Macro2.0.txt                     edit.cpp.1803.diff                     editor.cpp.1803.diff                     far.dep.1803.diff                     far.dsp.1803.diff                     Far.mak.1803.diff                     fn.hpp.1803.diff                     macro.hpp.1803.diff                     main.cpp.1803.diff                     mkfar.mak.1803.diff Основан на патче  : 1802 Дополнение        :
BugZ#1096 - shiftins в залоченном файле - двойное выделение.
  FAR 1799 без плагинов.
  в клипбоарде что-то есть.
  [ ] Expand Tabs to spaces
  [x] Persistent blocks
  [ ] Del removes blocks
  [ ] Auto indent
  [ ] Save file position
  [ ] Save bookmarks
  [x] Autodetect character table
  [x] Cursor beyond end of line
  [x] Lock editing of read-only files
  [ ] Warn when opening read-only files
  8   Tab size
  f4 на r/o файле.
  shiftins shiftright появился блок. идем в другое место файла и выделяем
  второй блок. первый остается.
  как я понимаю, в логике shiftins перед проверкой на то, что файл залочен,
  делаются какие-то внутренние операции с блоками от которых остаются хвосты.

Для "активных" действий проверим флаг FEDITOR_LOCKMODE.


В основные файлы, которые будут меняться (но чтобы не мешать основной
работе...) добавлен код от второй версии макро-движка. Ограничен
"#if defined(MACRODRIVE2)".

Сам движок находится в файле macro2.cpp, который по определенным
соображениям не выкладываю (не то, чтобы сыро, но... как будет
полностью готов, тогда и...)

Так же, файлы syntax.cpp и syntax.hpp - просто заглужки. Опять же
для того, чтобы по минимуму ворочить сорцы.

Хмм.. желательно проект перекомпилить!

+++ far/editor.cpp
@@ -6,10 +6,12 @@ editor.cpp
 
 */
 
-/* Revision: 1.249 29.05.2004 $ */
+/* Revision: 1.250 15.06.2004 $ */
 
 /*
 Modify:
+  15.06.2004 SVS
+    - BugZ#1096 - shiftins в залоченном файле - двойное выделение.
   29.05.2004 SVS
     - BugZ#1048 - Сдвигается влево отображение в редакторе по CtrlShiftEnd
       Строку "CurLine->EditLine.ObjWidth=X2-X1;" нужно было ставить до
@@ -2416,6 +2418,8 @@ int Editor::ProcessKey(int Key)
     case KEY_CTRLP:
     case KEY_CTRLM:
     {
+      if (Flags.Check(FEDITOR_LOCKMODE))
+        return TRUE;
       if (BlockStart!=NULL || VBlockStart!=NULL)
       {
         int SelStart,SelEnd;
@@ -2458,6 +2462,8 @@ int Editor::ProcessKey(int Key)
     }
     case KEY_CTRLD:
     {
+      if (Flags.Check(FEDITOR_LOCKMODE))
+        return TRUE;
       Flags.Clear(FEDITOR_MARKINGVBLOCK|FEDITOR_MARKINGBLOCK);
       DeleteBlock();
       Show();
@@ -2467,6 +2473,8 @@ int Editor::ProcessKey(int Key)
     case KEY_CTRLV:
     case KEY_SHIFTINS: case KEY_SHIFTNUMPAD0:
     {
+      if (Flags.Check(FEDITOR_LOCKMODE))
+        return TRUE;
       /* $ 10.04.2001 SVS
          Забыли Pasting выставить :-(
       */
